requirements.txt
matplotlib==3.5.3
numpy==1.23.5
Pillow==11.0.0
torch==2.5.1+cu124
torchvision==0.20.1+cu124
